+++ src/main/java/greedy/greedybot/application/googleform/dto/client/GoogleFormInformationResponse.java
@@ -1,12 +1,16 @@
-package greedy.greedybot.application.googleform.dto;
+package greedy.greedybot.application.googleform.dto.client;
 
 public record GoogleFormInformationResponse(
         Info info
 ) {
-    public record Info(
+    record Info(
             String title,
             String description,
             String documentTitle
     ) {
+    }
+
+    public String title() {
+        return info.title();
     }
 }
