+++ src/main/java/greedy/greedybot/application/googleform/dto/client/GoogleFormResponsesData.java
@@ -1,4 +1,4 @@
-package greedy.greedybot.application.googleform.dto;
+package greedy.greedybot.application.googleform.dto.client;
 
 import java.util.List;
 
@@ -6,7 +6,7 @@ public record GoogleFormResponsesData(
         List<Response> responses
 ) {
 
-    public record Response(
+    record Response(
             String responseId
     ) {
     }
